Create Accounts, Account, and Form components
created Form component so user can type the username and tag of the players
valorant account, the Accounts component will display all of the players
that are stored in the database on the screen. The Accounts component will
get this information from our Account component which will be displaying
one account at a time.