Return a response from createAccount funtion that postAccount action can use to populate redux store, create both a 201 res and 409 res.
createAccount REST endpoint didn't include a response which would hold up the
await functionality in the postAccounnt function. This was fixed by making
createAccount a asynchrounous function and letting it return a response
if it worked correctly(200) and if there was an error(400). Redux store
also has to be correctly populated which will allow us to fill the accounts
on the left side of the screen in the future.