Retrieve accounts from account state and use map function to pass them to Account component for more detailed information to be displayed
Previously we had text being displayed, now we are takinge every account
in the store using map function and passing them one at a time to the
Account component which will display it on the page with its full info
in the future. I also made sure we are using a grid so the layout of the accounts
is displayed in a neat looking grid. We also imported the circular progress
loader so when the account list is empty we see a loading sign.