Modify date so every account displayed has the correct time it was last received
Previously the entries in the database were missing found at value in its
schema. This resulted in all of the accounts being defaulted to the
same date and would only be unique at start up of the backend. To fix This
I removed the default value in the accountmodel of mongoose and just made
a new variable with the latest date value in the backend post function.
This will ensure every account will have the correct date of creation.